python/RirIpv4Allocated_cidr/src/requirements.txt
(py_psycopg2) $ pip freeze
mypy==1.10.1
mypy-extensions==1.0.0
psycopg2-binary==2.9.9
tomli==2.0.1
types-psycopg2==2.9.21.20240417
typing_extensions==4.12.2

